requirements.txt
aeneas==1.7.3.0
av==11.0.0
blobfile==2.1.1
cog==0.9.8
dac==0.4.2
datasets==2.16.0
demucs==4.0.1
dora_search==0.1.12
einops==0.8.0
ffmpeg_python==0.2.0
flashy==0.0.2
Flask==3.0.3
gradio==4.31.4
hear21passt==0.0.26
huggingface_hub==0.22.2
hydra-core==1.3.2
iou==0.1.0
ipython==8.12.3
julius==0.2.7
laion_clap==1.1.4
librosa==0.10.2.post1
lws==1.2.8
mpi4py_mpich==3.1.2
nltk==3.8.1
num2words==0.5.13
numpy==1.26.4
omegaconf==2.3.0
opencv_python==4.5.5.64
phonemizer==3.2.1
Pillow==10.3.0
protobuf==5.26.1
pytest==8.2.1
scipy==1.13.0
setuptools==69.5.1
soundfile==0.12.1
spacy==3.5.2
submitit==1.5.1
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.11.1
torchvision==0.15.2
tqdm==4.66.4
transformers==4.39.3
treetable==0.2.5
typing_extensions==4.11.0
whisperx==3.1.1
xformers==0.0.22
